+++ src/Http/Requests/CreateOrderRequest.php
@@ -13,11 +13,16 @@ use Spatie\LaravelData\Mappers\SnakeCaseMapper;
 #[MapOutputName(SnakeCaseMapper::class)]
 class CreateOrderRequest extends RevolutRequest
 {
+    /**
+     * @param  array<string, mixed>|null  $metadata
+     */
     public function __construct(
         public int $amount,
         public string $currency,
         public string $captureMode = 'automatic',
         public ?string $customerId = null,
         public ?string $redirectUrl = null,
+        public ?string $description = null,
+        public ?array $metadata = null,
     ) {}
 }